feat(sso):generate token for account

+++ usefulApps/sso/models/models.go
@@ -1,7 +1,7 @@
 package models
 
 type Account struct {
-	Id       int64  `gorm:"not null" json:"account_id"`
+	Id       int    `gorm:"not null" json:"account_id"`
 	Login    string `gorm:"not null" json:"account_login"`
 	Password string `gorm:"not null" json:"account_password"`
 }
@@ -11,7 +11,7 @@ func (Account) TableName() string {
 }
 
 type Token struct {
-	Id         int64  `gorm:"not null" json:"token_id"`
+	Id         int    `gorm:"not null" json:"token_id"`
 	TokenValue string `gorm:"not null" json:"token_value"`
 }
 
@@ -20,9 +20,9 @@ func (Token) TableName() string {
 }
 
 type AccountToken struct {
-	Id             int64 `gorm:"not null" json:"account_token_id"`
-	AccountReferId int64
-	TokenReferId   int64
+	Id             int `gorm:"not null" json:"account_token_id"`
+	AccountReferId int
+	TokenReferId   int
 }
 
 func (AccountToken) TableName() string {
